+++ PWGHF/TableProducer/candidateSelectorToXiPi.cxx
@@ -186,7 +186,7 @@ struct HfCandidateSelectorToXiPi {
       auto trackPiFromLam = trackV0NegDau;
       auto trackPrFromLam = trackV0PosDau;
 
-      int8_t signDecay = candidate.signDecay(); // sign of pi <- cascade
+      int signDecay = candidate.signDecay(); // sign of pi <- cascade
 
       if (signDecay > 0) {
         trackPiFromLam = trackV0PosDau;
